Devices: warnings (clang)
svn:sync-xref-src-repo-rev: r110129

+++ src/VBox/Devices/Graphics/DevVGA-SVGA.cpp
@@ -1,4 +1,4 @@
-/* $Id: DevVGA-SVGA.cpp 63421 2016-08-13 17:12:29Z knut.osmundsen@oracle.com $ */
+/* $Id: DevVGA-SVGA.cpp 63478 2016-08-15 14:04:10Z knut.osmundsen@oracle.com $ */
 /** @file
  * VMWare SVGA device.
  *
@@ -3105,8 +3105,8 @@ static DECLCALLBACK(int) vmsvgaFIFOLoop(PPDMDEVINS pDevIns, PPDMTHREAD pThread)
 
             default:
 # ifdef VBOX_WITH_VMSVGA3D
-                if (    enmCmdId >= SVGA_3D_CMD_BASE
-                    &&  enmCmdId <  SVGA_3D_CMD_MAX)
+                if (    (int)enmCmdId >= SVGA_3D_CMD_BASE
+                    &&  (int)enmCmdId <  SVGA_3D_CMD_MAX)
                 {
                     /* All 3d commands start with a common header, which defines the size of the command. */
                     SVGA3dCmdHeader *pHdr;
